fix: finish quota client review
