+++ cmd/drive9/cli/admin.go
@@ -363,134 +363,6 @@ func printAdminTenantTable(tenants []client.AdminTenant, includeQuota bool) erro
 	return w.Flush()
 }
 
-func quotaTableHeader(includeKind bool) string {
-	columns := []string{"TENANT_ID", "STATUS"}
-	if includeKind {
-		columns = append(columns, "KIND")
-	}
-	columns = append(columns,
-		"MAX_STORAGE",
-		"MAX_FILE_SIZE",
-		"MAX_FILE_COUNT",
-		"MAX_MEDIA_LLM_FILES",
-		"MAX_VIDEO_LLM_FILES",
-		"SPENDING_LIMIT",
-		"STORAGE_USED",
-		"RESERVED",
-		"FILE_COUNT",
-		"MEDIA_FILE_COUNT",
-		"VIDEO_FILE_COUNT",
-	)
-	return strings.Join(columns, "\t")
-}
-
-func quotaTableRow(tenantID, status, kind string, quota *client.AdminTenantQuota, includeKind bool) string {
-	values := []string{tenantID, status}
-	if includeKind {
-		values = append(values, kind)
-	}
-	values = append(values,
-		adminQuotaMaxStorage(quota),
-		adminQuotaMaxFileSize(quota),
-		adminQuotaMaxFileCount(quota),
-		adminQuotaMaxMediaLLMFiles(quota),
-		adminQuotaMaxVideoLLMFiles(quota),
-		adminQuotaSpendingLimit(quota),
-		adminQuotaStorageUsed(quota),
-		adminQuotaReserved(quota),
-		adminQuotaFileCount(quota),
-		adminQuotaMediaFileCount(quota),
-		adminQuotaVideoFileCount(quota),
-	)
-	return strings.Join(values, "\t")
-}
-
-func adminQuotaMaxStorage(quota *client.AdminTenantQuota) string {
-	if quota == nil {
-		return "-"
-	}
-	return fmt.Sprintf("%d Mi", quota.Config.MaxStorageSize)
-}
-
-func adminQuotaMaxFileSize(quota *client.AdminTenantQuota) string {
-	if quota == nil {
-		return "-"
-	}
-	return fmt.Sprintf("%d Mi", quota.Config.MaxFileSize)
-}
-
-func adminQuotaMaxFileCount(quota *client.AdminTenantQuota) string {
-	if quota == nil {
-		return "-"
-	}
-	if quota.Config.MaxFileCount == 0 {
-		return "unlimited"
-	}
-	return fmt.Sprintf("%d", quota.Config.MaxFileCount)
-}
-
-func adminQuotaMaxMediaLLMFiles(quota *client.AdminTenantQuota) string {
-	if quota == nil {
-		return "-"
-	}
-	if quota.Config.MaxMediaLLMFiles == 0 {
-		return "unlimited"
-	}
-	return fmt.Sprintf("%d", quota.Config.MaxMediaLLMFiles)
-}
-
-func adminQuotaMaxVideoLLMFiles(quota *client.AdminTenantQuota) string {
-	if quota == nil {
-		return "-"
-	}
-	if quota.Config.MaxVideoLLMFiles == 0 {
-		return "unlimited"
-	}
-	return fmt.Sprintf("%d", quota.Config.MaxVideoLLMFiles)
-}
-
-func adminQuotaSpendingLimit(quota *client.AdminTenantQuota) string {
-	if quota == nil || quota.Config.TiDBCloudSpendingLimit == nil {
-		return "-"
-	}
-	return fmt.Sprintf("%d", *quota.Config.TiDBCloudSpendingLimit)
-}
-
-func adminQuotaStorageUsed(quota *client.AdminTenantQuota) string {
-	if quota == nil {
-		return "-"
-	}
-	return formatBytes(quota.Usage.StorageBytes)
-}
-
-func adminQuotaReserved(quota *client.AdminTenantQuota) string {
-	if quota == nil {
-		return "-"
-	}
-	return formatBytes(quota.Usage.ReservedBytes)
-}
-
-func adminQuotaFileCount(quota *client.AdminTenantQuota) string {
-	if quota == nil {
-		return "-"
-	}
-	return fmt.Sprintf("%d", quota.Usage.FileCount)
-}
-
-func adminQuotaMediaFileCount(quota *client.AdminTenantQuota) string {
-	if quota == nil {
-		return "-"
-	}
-	return fmt.Sprintf("%d", quota.Usage.MediaFileCount)
-}
-
-func adminQuotaVideoFileCount(quota *client.AdminTenantQuota) string {
-	if quota == nil {
-		return "-"
-	}
-	return fmt.Sprintf("%d", quota.Usage.VideoFileCount)
-}
-
 func adminTenantDelete(args []string) error {
 	tenantID, server, publicKey, privateKey, asJSON, err := parseAdminTenantIDCommand(args, adminTenantDeleteUsage())
 	if err != nil {
@@ -858,6 +730,8 @@ flags:
   --max-storage-size Mi            set-quota: max confirmed+reserved storage size in Mi
   --max-file-size Mi               set-quota: max single file size in Mi
   --max-file-count N               set-quota: max confirmed file count; 0 means unlimited
+  --max-media-llm-files N          set-quota: max media LLM extract files; non-negative; 0 means unlimited; omit to keep current; requires media extract config
+  --max-video-llm-files N          set-quota: max video LLM extract files; non-negative; 0 means unlimited; omit to keep current; requires video extract config
   --tidbcloud-spending-limit N     set-quota: TiDB Cloud Cluster Spending Limit
   --json                           output result as JSON when supported
