+++ cmd/drive9/visual_help.go
@@ -489,8 +489,8 @@ func drive9VisualHelpCommands() []visualHelpCommand {
 				{Name: "--max-storage-size Mi", Desc: "set-quota: max confirmed+reserved storage size in Mi"},
 				{Name: "--max-file-size Mi", Desc: "set-quota: max single file size in Mi"},
 				{Name: "--max-file-count N", Desc: "set-quota: max confirmed file count; 0 means unlimited"},
-				{Name: "--max-media-llm-files N", Desc: "set-quota: max media LLM extract files; must be non-negative; requires media extract config"},
-				{Name: "--max-video-llm-files N", Desc: "set-quota: max video LLM extract files; must be non-negative; requires video extract config"},
+				{Name: "--max-media-llm-files N", Desc: "set-quota: max media LLM extract files; non-negative; 0 means unlimited; omit to keep current; requires media extract config"},
+				{Name: "--max-video-llm-files N", Desc: "set-quota: max video LLM extract files; non-negative; 0 means unlimited; omit to keep current; requires video extract config"},
 				{Name: "--tidbcloud-spending-limit N", Desc: "tenant create/set-quota TiDB Cloud Cluster Spending Limit"},
 				{Name: "--json", Desc: "output result as JSON"},
 			},
